+++ app/src/main/java/com/example/slacks_lottoevent/ManageMyEventsFragment.java
@@ -5,7 +5,9 @@ import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Button;
 import android.widget.ListView;
+import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -18,25 +20,92 @@ import com.google.firebase.firestore.EventListener;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.FirebaseFirestoreException;
 import com.google.firebase.firestore.QuerySnapshot;
-import com.google.zxing.common.BitMatrix;
 
 import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.Map;
 
 /**
  * This activity is responsible for managing the events created by the organizer.
  */
-public class ManageMyEventsFragment extends Fragment {
+public class ManageMyEventsFragment extends Fragment implements AddFacilityFragment.AddFacilityDialogListener{
 
     private OrganzierEventArrayAdapter organzierEventArrayAdapter;
     private ListView myEventsListView;
     private FragmentManageMyEventsBinding binding;
+    private TextView facilityCreated;
+    private Facility existingFacility;
+    private Button createFacilitiesButton;
 
     private FirebaseFirestore db;
     private CollectionReference eventsRef;
     private ArrayList<Event> eventList;
+    private CollectionReference facilitiesRef;
 
 //    organzierEventArrayAdapter = new void OrganzierEventArrayAdapter(getContext(), eventList);
+
+    @Override
+    public void addFacility(Facility facility) {
+        // Retrieve the facility name and set it to display on the screen
+        String facilityName = facility.getFacilityName();
+        facilityCreated.setText(facilityName);
+        existingFacility = facility;
+
+        Map<String, Object> facilityData = new HashMap<>();
+        facilityData.put("name", facilityName);
+        // Add other attributes as needed, like location, type, etc.
+        facilityData.put("streetAddress1", facility.getStreetAddress1());
+        facilityData.put("streetAddress2", facility.getStreetAddress2());
+        facilityData.put("city", facility.getCity());
+        facilityData.put("province", facility.getProvince());
+        facilityData.put("country", facility.getCountry());
+        facilityData.put("postalCode", facility.getProvince());
+
+        // Add the facility data to Firestore
+        facilitiesRef.add(facilityData)
+                .addOnSuccessListener(documentReference -> {
+                    Log.d("addFacility", "Facility added with ID: " + documentReference.getId());
+                    existingFacility.setDocumentId(documentReference.getId()); // Set document ID here
+                })
+                .addOnFailureListener(e -> Log.w("addFacility", "Error adding facility", e));
+
+        // Hide the create facility button
+        createFacilitiesButton.setVisibility(View.GONE);
+
+    }
+
+    @Override
+    public void updateFacility() {
+        String facilityName = existingFacility.getFacilityName();
+        facilityCreated.setText(facilityName);
+
+        if (existingFacility == null) {
+            Log.w("updateFacility", "No facility selected to update");
+            return;
+        }
+
+        // Retrieve the document ID for the existing facility
+        String documentId = existingFacility.getDocumentId();
+        if (documentId == null || documentId.isEmpty()) {
+            Log.w("updateFacility", "No document ID provided for update");
+            return;
+        }
+
+        // Prepare the updated facility data
+        Map<String, Object> facilityData = new HashMap<>();
+        facilityData.put("name", existingFacility.getFacilityName());
+        facilityData.put("streetAddress1", existingFacility.getStreetAddress1());
+        facilityData.put("streetAddress2", existingFacility.getStreetAddress2());
+        facilityData.put("city", existingFacility.getCity());
+        facilityData.put("province", existingFacility.getProvince());
+        facilityData.put("country", existingFacility.getCountry());
+        facilityData.put("postalCode", existingFacility.getPostalCode());
+
+        // Update the document in Firestore
+        facilitiesRef.document(documentId).update(facilityData)
+                .addOnSuccessListener(aVoid -> Log.d("updateFacility", "Facility updated successfully"))
+                .addOnFailureListener(e -> Log.w("updateFacility", "Error updating facility", e));
+    }
 
     @Override
     public View onCreateView(
@@ -50,13 +119,18 @@ public class ManageMyEventsFragment extends Fragment {
     public void onViewCreated(@NonNull View view, Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
 
+
         eventList = new ArrayList<>(); // Initialize the event list
         db = FirebaseFirestore.getInstance();
         eventsRef = db.collection("events"); // Reference to events collection
+        facilitiesRef = db.collection("facilities");
 
         myEventsListView = binding.myEventsListView;
         organzierEventArrayAdapter = new OrganzierEventArrayAdapter(getContext(), eventList);
         myEventsListView.setAdapter(organzierEventArrayAdapter);
+
+        createFacilitiesButton = view.findViewById(R.id.create_facility_button);
+        facilityCreated = view.findViewById(R.id.facility_created);
 
         eventsRef.addSnapshotListener(new EventListener<QuerySnapshot>() {
             @Override
@@ -127,6 +201,36 @@ public class ManageMyEventsFragment extends Fragment {
                     }
                 }
             }
+        });
+        facilitiesRef.addSnapshotListener((querySnapshot, e) -> {
+            if (e != null) {
+                Log.w("Firestore", "Listen failed.", e);
+                return;
+            }
+
+            if (querySnapshot != null && !querySnapshot.isEmpty()) {
+                DocumentSnapshot document = querySnapshot.getDocuments().get(0);
+                String facilityName = document.getString("name");
+                if (facilityName != null && !facilityName.isEmpty()) {
+                    facilityCreated.setText(facilityName);
+                    facilityCreated.setVisibility(View.VISIBLE);
+                } else {
+                    facilityCreated.setVisibility(View.GONE);
+                }
+            } else {
+                facilityCreated.setVisibility(View.GONE);
+                createFacilitiesButton.setVisibility(View.VISIBLE);
+            }
+        });
+
+        createFacilitiesButton.setOnClickListener(new Button.OnClickListener(){
+            public void onClick(View v){
+                new AddFacilityFragment().show(getParentFragmentManager(), "Add Facility");
+            }
+        });
+
+        facilityCreated.setOnClickListener(v -> {
+            new AddFacilityFragment(existingFacility, true).show(getParentFragmentManager(), "Edit Facility");
         });
 
     }
